Make Generic OAuth available to JS devices
Cause why not? It covers most of the use cases, and fancier
use cases can still use custom_oauth/custom JS code.
