+++ lib/modules/base_generic.js
@@ -12,46 +12,6 @@
 const Tp = require('thingpedia');
 const Utils = require('./utils');
 
-function makeGenericOAuth(kind, ast, devclass) {
-    function OAuthCallback(engine, accessToken, refreshToken, extraData) {
-        var obj = { kind: kind,
-                    accessToken: accessToken,
-                    refreshToken: refreshToken };
-        for (let name in extraData) {
-            if (name === 'access_token' || name === 'refresh_token' || name === 'expires_in')
-                continue;
-            obj[name] = extraData[name];
-        }
-
-        if (ast.auth.get_profile) {
-            var auth = 'Bearer ' + accessToken;
-            return Tp.Helpers.Http.get(ast.auth.get_profile, { auth: auth,
-                                                               accept: 'application/json' })
-                .then((response) => {
-                    var profile = JSON.parse(response);
-
-                    ast.auth.profile.forEach((p) => {
-                        obj[p] = profile[p];
-                    });
-
-                    return engine.devices.loadOneDevice(obj, true);
-                });
-        } else {
-            return engine.devices.loadOneDevice(obj, true);
-        }
-    }
-
-    var runOAuth2 = Tp.Helpers.OAuth2({ kind: kind,
-                                     client_id: ast.auth.client_id,
-                                     client_secret: ast.auth.client_secret,
-                                     authorize: ast.auth.authorize,
-                                     get_access_token: ast.auth.get_access_token,
-                                     set_state: ast.auth.set_state,
-                                     callback: OAuthCallback });
-    runOAuth2.install(devclass.prototype);
-    devclass.runOAuth2 = runOAuth2;
-}
-
 module.exports = class BaseGenericModule {
     constructor(kind, ast) {
         this._id = kind;
@@ -92,7 +52,7 @@ module.exports = class BaseGenericModule {
             }
         };
         if (ast.auth.type === 'oauth2')
-            makeGenericOAuth(kind, ast, this._loaded);
+            Utils.makeGenericOAuth(kind, ast, this._loaded);
 
         this._loaded.metadata = ast;
     }
