+++ lib/modules/base_js.js
@@ -18,6 +18,8 @@ const Module = require('module');
 
 const Tp = require('thingpedia');
 
+const { makeGenericOAuth } = require('./utils');
+
 // shared code between v1 and v2 modules
 
 function resolve(mainModule) {
@@ -96,7 +98,9 @@ module.exports = class BaseJavascriptModule {
     _completeLoading(deviceClass) {
         deviceClass.metadata = this._manifest;
 
-        if (deviceClass.runOAuth2 && deviceClass.runOAuth2.install)
+        if (this._manifest.auth.type === 'oauth2' && !deviceClass.runOAuth2)
+            makeGenericOAuth(this._id, this._manifest, deviceClass);
+        else if (deviceClass.runOAuth2 && deviceClass.runOAuth2.install)
             deviceClass.runOAuth2.install(deviceClass.prototype);
     }
