+++ lib/modules/utils.js
@@ -9,6 +9,7 @@
 // See LICENSE for details
 "use strict";
 
+const Tp = require('thingpedia');
 const ThingTalk = require('thingtalk');
 const Units = ThingTalk.Units;
 
@@ -25,6 +26,50 @@ function measureToString(value, precision, unit) {
 }
 
 module.exports = {
+    makeGenericOAuth(kind, ast, devclass) {
+        function OAuthCallback(engine, accessToken, refreshToken, extraData) {
+            var obj = { kind: kind,
+                        accessToken: accessToken,
+                        refreshToken: refreshToken };
+            for (let name in extraData) {
+                if (name === 'access_token' || name === 'refresh_token' || name === 'expires_in')
+                    continue;
+                obj[name] = extraData[name];
+            }
+
+            if (ast.auth.get_profile) {
+                var auth = 'Bearer ' + accessToken;
+                return Tp.Helpers.Http.get(ast.auth.get_profile, { auth: auth,
+                                                                   accept: 'application/json' })
+                    .then((response) => {
+                        var profile = JSON.parse(response);
+
+                        if (ast.auth.profile) {
+                            ast.auth.profile.forEach((p) => {
+                                obj[p] = profile[p];
+                            });
+                        } else {
+                            obj.profile = profile;
+                        }
+
+                        return engine.devices.loadOneDevice(obj, true);
+                    });
+            } else {
+                return engine.devices.loadOneDevice(obj, true);
+            }
+        }
+
+        var runOAuth2 = Tp.Helpers.OAuth2({ kind: kind,
+                                         client_id: ast.auth.client_id,
+                                         client_secret: ast.auth.client_secret,
+                                         authorize: ast.auth.authorize,
+                                         get_access_token: ast.auth.get_access_token,
+                                         set_state: !!ast.auth.set_state,
+                                         callback: OAuthCallback });
+        runOAuth2.install(devclass.prototype);
+        devclass.runOAuth2 = runOAuth2;
+    },
+
     formatString(url, deviceParams, functionParams) {
         return url.replace(/\$(?:\$|([a-zA-Z0-9_]+(?![a-zA-Z0-9_]))|{([a-zA-Z0-9_]+)(?::(%|[a-zA-Z-]+))?})/g, (match, param1, param2, opt) => {
             if (match === '$$')
